backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
requests
pillow
openai
python-dotenv
aiofiles
fal-client
